Import the ffmpg-devel package.
FFmpeg is the first complete and free Internet Live Audio and Video
Broadcasting solution. FFMpeg aims at being the command line tool to
handle audio and video. It is a "three-in-one" solution.

It is the last version of ffmpeg ( we need it to build the last vlc : I will commit it sooner )
The package is based under the multimedia/ffmpeg.
I have corrected the patches ( in order to match the new sources) and add some options for the package ( faad faac a52)

The package runs well but doesn't compile with the mmx options ( I don't understand the asm used into gcc so I will have somme difficulties to solve the problem : please check the TODO).