nss is the netscape tls library that ships with mozilla, but packaged up by itself.  nss is like gnutls and openssl.
the motivation for this package is borne out of problems with gnutls
and gaim.

http://mail-index.netbsd.org/tech-pkg/2004/12/22/0001.html

http://mail-index.netbsd.org/tech-pkg/2004/12/22/0004.html

http://mail-index.netbsd.org/tech-pkg/2005/01/04/0004.html

in addition, i know of someone else with gaim-encryption work in
progress that requires nss to work.

this package (like nspr) is heavily inspired by the freebsd port's
construction, with help from pkgsrc/www/mozilla/

this package does not currently build.  the error message is in
TODO.  i'm going to ask for help shortly.