Gaim plug-in for libotr.  This enables confidential (with PFS), authenticated, and deniable messaging over any IM transport.
Note that the distfile is unfetchable, because I just converted it to
use autoconf/automake (and sent patches to the upstream maintainer)
and the new version has not yet been released.  I expect that within a
few days.  A patch against 2.0.1 is at
http://lists.cypherpunks.ca/pipermail/otr-dev/2005-February/000353.html
for those wanting to generate a 2.0.2 tarball.