Add MESSAGE to suggest installing sysutils/cdrdao package for rezound to burn audio files to CD.
Add patch-af to fix problem with detecting bison.

Get rid of PKGNAME because lets keep this name lowercase
(and include "beta" in name).

Fix MASTER_SITES to use sourceforge mirrors.

Improve COMMENT.

Get rid of BUILD_DEPENDS for bison and cdrdao.
Add comment that needs real bison and not yacc.
Add "yacc" to USE_GNU_TOOLS. This will be broken
on some platforms. I understand the new "tools" system
for pkgsrc will have a choice for "bison".

Support the pkgsrc DEVOSSSOUND. Also patch-ad uses this.

Comment out the LDFLAGS for -lcompat. (This package
does not have a maintainer, sorry if this breaks for someone.)

Get rid of setting "YACC". It is overridden on some systems
and not seen anyways.

Add a build time option for "alsa".

Include buildlinks for pthread and ossaudio.

Add locale files to PLIST.

Todo: see the todo!