initial import of png2ico
DESCR:

Converts PNG files to Windows icon resource files. If you're looking for a
program to create a favicon.ico for your website, look no further. If you
need instructions or don't even know what a favicon is, check out my short
tutorial on how to create and install a favicon.ico.

The program is extremely simple to use. To create a favicon.ico that contains
a logo in the resolutions 16x16 and 32x32 (an icon can contain multiple
images of different sizes, but a single 16x16 image is enough for a favicon),
you would use a command like the following:

     png2ico favicon.ico logo16x16.png logo32x32.png