nrpe is used to execute Nagios plugins on remote hosts and report the results to the main Nagios server. From the Nagios homepage:
Allows you to execute "local" plugins (like check_disk, check_procs, etc.) on
remote hosts. The check_nrpe plugin is called from Nagios and actually makes
the plugin requests to the remote host. Requires that nrpe be running on the
remote host (either as a standalone daemon or as a service under inetd).

nrpe2 is incompatible with nrpe version 1, if you are running this version
see net-mgmt/nrpe. Both ports can be installed in parallel but share the same
configuration file by default. This is easily changed in the nrpe startup
file.

WWW: http://www.nagios.org/