Suppose you have a large piece of text - typically, say, a web page or a mail message. And now suppose you've done some kind of full-text search on that text for a bunch of keywords, and you want to display the context in which you found the keywords inside the body of the text.
A simple-minded way to do that would be just to get the two words either
side of each keyword. But hey, don't be too simple minded, because you've
got to make sure that the list doesn't overlap. If you have

    the quick brown fox jumped over the lazy dog

and you extract two words either side of ``fox'', ``jumped'' and ``dog'',
you really don't want to end up with

    quick brown fox jumped over brown fox jumped over the the lazy dog

so you need a small amount of smarts. This module has a small amount of smarts.