initial import for libvc
This vCard library was written using `flex' and `bison'.  The library
handles scanning a vCard file, parsing the vCard file into a data
storage structure, manipulating the data storage structure, and
writing back to file.  It was originally written for `rolo', but
has been split-off as a generic library for general use.