initial import of heme
Heme is a fast and portable console hex editor for unix systems. It has undo
support (number of undo operations is only limited by available memory),
ability to fill a range of addresses with the specified byte, ability to
search for a single byte or character string.

Offsets can be given in hexadecimal, octal or decimal. There are two editing
modes: hex and ascii. In hex mode you type hexadecimal digits and in ascii
mode you type characters. In the ascii mode cursor is automatically moved to
the next byte after typing a character (this behaviour is configurable for
hex mode).