Mednafen is a portable, utilizing OpenGL and SDL, command-line-driven multi-system emulator with many advanced features. The Atari Lynx, GameBoy, GameBoy Color, GameBoy Advance, NES, PC Engine(TurboGrafx 16), and SuperGrafx are emulated.
Mednafen has the ability to remap hotkey functions and virtual system inputs
to a keyboard, a joystick, or both simultaneously. Save states are supported,
as is real-time game rewinding. Screen snapshots may be taken at the press of
a button, and are saved in the popular PNG file format.

Mednafen is distributed under the terms of the GPL.

Due to the threaded model of emulation used in Mednafen, and limitations of
SDL, a joystick is preferred over a keyboard to play games, as the joystick
will have slightly less latency, although the latency differences may not be
perceptible to most people.