Initial import of FreeBSD hcidump utility for NetBSD. This was originally an older version of bluez-hcidump which is already in wip, but the bluez version has evolved out of sight and its too much work to make it compile.
HCIdump reads raw HCI data coming from and going to a Bluetooth device and
prints to screen commands, events and data in a human-readable form.
Optionally, the dump can be written to a file rather than parsed, and the dump
file can be parsed in a subsequent moment.

Incidentally you probably wont be able to get this to compile either as you
wont have the header files that I havent finished writing although they can 
be found at "http://homepages.rya-online.net/plunky/netbt.html" along with a
preliminary start on a bluetooth protocol stack for NetBSD. I just want to make
this available for now.