Haddock is a tool for automatically generating documentation from annotated Haskell source code. It is primarily intended for documenting libraries, but it should be useful for any kind of Haskell code.
Like other systems, Haddock lets you write documentation annotations
next to the definitions of functions and types in the source code, in a
syntax that is easy on the eye when writing the source code (no
heavyweight mark-up). The documentation generated by Haddock is fully
hyperlinked - click on a type name in a type signature to go straight to
the definition, and documentation, for that type.