initial import of Rhapsody ==========================
Rhapsody is an IRC client for Unix operating systems intended to be displayed
on a text console. It is fast, portable and easy to use, yet it is full
featured. It provides an intuitive, menu-driven user interface, and is ideal
for beginner to intermediate users.


TODO: help doesn't works, requires patches in sources, but it's a beta version,
and this would be fixed soon (i hope :P)