Import Mozilla Seamonkey 1.0 (gtk1 version).
The SeaMonkey project is a community effort to deliver production-quality
releases of code derived from the application formerly known as "Mozilla
Application Suite". Whereas the main focus of the Mozilla Foundation is on
Mozilla Firefox and Mozilla Thunderbird, our group of dedicated volunteers
works to ensure that you can have "everything but the kitchen sink" and have
it stable enough for corporate use.

This package uses GTK+-1.x widgets.