* Initial import of fusefs-sshfs
This is a filesystem client based on the SSH File Transfer Protocol.
Since most SSH servers already support this protocol it is very easy
to set up: i.e. on the server side there's nothing to do.  On the
client side mounting the filesystem is as easy as logging into the
server with ssh.

The idea of sshfs was taken from the SSHFS filesystem distributed with
LUFS, which I found very useful.  There were some limitations of that
codebase, so I rewrote it.  Features of this implementation are:

  - Based on FUSE (the best userspace filesystem framework for Linux ;)
  - Multithreading: more than one request can be on it's way to the
    server
  - Allowing large reads (max 64k)
  - Caching directory contents
  - Reconnect on failure