Initial import of the epsg-6.11 package.
EPSG Geodetic Parameters

The OGP Surveying and Positioning Committee, through its geodesy
sub-committee, maintains and publishes a dataset of parameters for
coordinate reference system and coordinate transformation description.
The EPSG Geodetic Parameter Dataset has been included as reference
data in UKOOA and SEG positioning data exchange formats, the GeoTIFF
interchange format for georeferenced raster imagery, the IHS Energy
Iris21, PPDM and POSC Epicentre data models.  The dataset is
distributed in an MS Access relational database and SQL script files.

This current release (Version 6.11) of the EPSG Geodetic Parameter
Dataset comprises of data updates to the v6.10 release of April 2006.
It also replaces interim releases v6.10.1 of May 2006 and v6.10.2 of
June 2006 which included updates to the SQL scripts accompanying the
v6.10 release.  The EPSG Geodetic Parameter Dataset, or EPSG dataset,
is maintained by the Geodesy Subcommittee of the Surveying and
Positioning Committee of the International Association of Oil and Gas
Producers (OGP).

This package includes the PostgreSQL and mySQL scripts for creating
the EPSG database.


EPSG LICENSE CONDITIONS

Use of the Data

1) All data pertinent to a specific coordinate reference system must
   be copied without modification and all related pages/records must
   be included;

2) All components of this data set pertinent to any given coordinate
   reference system must be distributed together (complete
   distribution of all components of the data set is preferred, but
   the OGP recognises the need for a more limited distribution);

3) The data may not be distributed for profit by any third party; and

4) The original source [OGP] must be acknowledged.

The user assumes the entire risk as to the accuracy and the use of
this data.  The data may be copied and distributed subject to the
following conditions:

INFORMATION PROVIDED IN THIS DOCUMENT IS PROVIDED "AS IS" WITHOUT
WARRANTY OF ANY KIND, EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT
LIMITED TO THE IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS
FOR A PARTICULAR PURPOSE.

With regard to (3) above, the data may be included within proprietary
applications distributed on a commercial basis when the commerciality
is based on application functionality and not on a value ascribed to
the freely-distributed EPSG dataset.