Only include libX11 and libXext buildlink3.mk files if X11_TYPE is "modular-xorg".
If X11_TYPE is "native" then set some configure environment variables
so pkg-config is bypassed. And if Xregion.h header is missing,
then copy this file from this files directory into the buildlink
includes directory.

Add a Xregion.h header from libX11 source to files directory.
It was not found in XFree86 in my tests on NetBSD 3.0.1.