Standalone GRASS 6 Drivers for GDAL and OGR 	===========================================
This package contains standalone drivers for GRASS raster and vector
files that can be built after GDAL has been built and installed as an
"autoload" driver.

This is particularly useful in resolving problems with GRASS depending
on GDAL, but GDAL with GRASS support depending on GRASS.  With this
package you can configure and install GDAL normally (--without-grass), then
build and install GRASS normally and finally build and install this driver.

See also:

  http://www.gdal.org/
  http://grass.itc.it/